+++ src/main/java/io/induct/yle/api/programs/domain/items/TvClip.java
@@ -1,9 +1,9 @@
-package io.induct.yle.api.programs.model.items;
+package io.induct.yle.api.programs.domain.items;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.induct.yle.api.YleId;
-import io.induct.yle.api.programs.model.Item;
+import io.induct.yle.api.programs.domain.Item;
 
 /**
  * @since 2015-05-30
