+++ src/test/java/io/induct/yle/api/programs/YleProgramsApiTest.java
@@ -11,10 +11,10 @@ import io.induct.yle.TestDependenciesModule;
 import io.induct.yle.YleApiTestingBase;
 import io.induct.yle.api.common.Language;
 import io.induct.yle.api.common.MediaObject;
-import io.induct.yle.api.programs.model.CuratedList;
-import io.induct.yle.api.programs.model.Item;
-import io.induct.yle.api.programs.model.items.Service;
-import io.induct.yle.api.programs.model.search.ItemSearch;
+import io.induct.yle.api.programs.domain.CuratedList;
+import io.induct.yle.api.programs.domain.Item;
+import io.induct.yle.api.programs.domain.items.Service;
+import io.induct.yle.api.programs.domain.search.ItemSearch;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.stubbing.OngoingStubbing;
